pyproject.toml
[project]
name = "nishuihan-dianshi-assistant"
version = "0.1.0"
description = "逆水寒殿试答题器"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "customtkinter==5.2.2",
    "httpx>=0.27.0",
    "pillow==10.4.0",
    "pytesseract==0.3.13",
    "requests==2.32.3"
]

[project.optional-dependencies]
build = [
    "pyinstaller>=6.0.0",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
